+++ leia-core/src/main/java/com/grookage/leia/core/ingestion/processors/ApproveSchemaProcessor.java
@@ -17,7 +17,7 @@
 package com.grookage.leia.core.ingestion.processors;
 
 import com.grookage.leia.core.exception.LeiaSchemaErrorCode;
-import com.grookage.leia.core.ingestion.utils.ContextUtils;
+import com.grookage.leia.core.ingestion.utils.SchemaUtils;
 import com.grookage.leia.models.exception.LeiaException;
 import com.grookage.leia.models.schema.SchemaDetails;
 import com.grookage.leia.models.schema.SchemaKey;
@@ -56,27 +56,10 @@ public class ApproveSchemaProcessor extends SchemaProcessor {
                     schemaKey.getSchemaName());
             throw LeiaException.error(LeiaSchemaErrorCode.NO_SCHEMA_FOUND);
         }
-        validateSchemaApproval(context, storedSchema);
+        SchemaUtils.validateSchemaApproval(context, storedSchema);
         addHistory(context, storedSchema);
         storedSchema.setSchemaState(SchemaState.APPROVED);
         getRepositorySupplier().get().update(storedSchema);
         context.addContext(SchemaDetails.class.getSimpleName(), storedSchema);
-    }
-
-    private void validateSchemaApproval(SchemaContext context, SchemaDetails storedSchema) {
-        final var schemaKey = storedSchema.getSchemaKey();
-        final String currentUserId = ContextUtils.getUserId(context);
-
-        // Check if current user has previously created or updated this schema
-        final var isCreatorOrUpdater = storedSchema.getHistories().stream()
-                .filter(history -> history.getSchemaEvent() == SchemaEvent.CREATE_SCHEMA ||
-                                   history.getSchemaEvent() == SchemaEvent.UPDATE_SCHEMA)
-                .anyMatch(history -> history.getConfigUpdaterId().equalsIgnoreCase(currentUserId));
-
-        if (isCreatorOrUpdater) {
-            log.error("User '{}' cannot approve schema '{}' because they previously created or updated it",
-                    ContextUtils.getUser(context), schemaKey.getReferenceId());
-            throw LeiaException.error(LeiaSchemaErrorCode.SCHEMA_APPROVAL_UNAUTHORIZED);
-        }
     }
 }
