+++ leia-core/src/main/java/com/grookage/leia/core/ingestion/utils/SchemaUtils.java
@@ -16,13 +16,20 @@
 
 package com.grookage.leia.core.ingestion.utils;
 
+import com.grookage.leia.core.exception.LeiaSchemaErrorCode;
+import com.grookage.leia.models.exception.LeiaException;
 import com.grookage.leia.models.schema.SchemaDetails;
+import com.grookage.leia.models.schema.engine.SchemaContext;
+import com.grookage.leia.models.schema.engine.SchemaEvent;
 import com.grookage.leia.models.schema.engine.SchemaState;
 import com.grookage.leia.models.schema.ingestion.CreateSchemaRequest;
 import lombok.experimental.UtilityClass;
+import lombok.extern.slf4j.Slf4j;
 
 @UtilityClass
+@Slf4j
 public class SchemaUtils {
+    private static final boolean LOCAL_ENV = Boolean.parseBoolean(System.getProperty("localEnv", "false"));
 
     public SchemaDetails toSchemaDetails(final CreateSchemaRequest createSchemaRequest) {
         return SchemaDetails.builder()
@@ -35,5 +42,26 @@ public class SchemaUtils {
                 .transformationTargets(createSchemaRequest.getTransformationTargets())
                 .tags(createSchemaRequest.getTags())
                 .build();
+    }
+
+    public void validateSchemaApproval(SchemaContext context, SchemaDetails storedSchema) {
+        if (LOCAL_ENV) {
+            // Skip approver validation in local environment
+            return;
+        }
+        final var schemaKey = storedSchema.getSchemaKey();
+        final String currentUserId = ContextUtils.getUserId(context);
+
+        // Check if current user has previously created or updated this schema
+        final var isCreatorOrUpdater = storedSchema.getHistories().stream()
+                .filter(history -> history.getSchemaEvent() == SchemaEvent.CREATE_SCHEMA ||
+                                   history.getSchemaEvent() == SchemaEvent.UPDATE_SCHEMA)
+                .anyMatch(history -> history.getConfigUpdaterId().equalsIgnoreCase(currentUserId));
+
+        if (isCreatorOrUpdater) {
+            log.error("User '{}' cannot approve schema '{}' because they previously created or updated it",
+                    ContextUtils.getUser(context), schemaKey.getReferenceId());
+            throw LeiaException.error(LeiaSchemaErrorCode.SCHEMA_APPROVAL_UNAUTHORIZED);
+        }
     }
 }
